fix: recognize breaking changes in version command (#101)
A naming mistake causes the accumulation of `breaking` changes to
yield `NaN` in case there is more than 0 breaking changes.  This was
caused by the reduction being initialized with a missnamed attribute
name (`breakingChanges` instead of `breaking`).

The `breaking` attribute was not set in the reduction, causing its
initial value to be `undefined`.  The increments for breaking changes
(`acc.breaking++`) therefore caused the evaluation of the expression
`undefined++`.  Therefore there were always `NaN` breaking changes
whenever there was more than one breaking change.